ugly update of spiking logic
Extremely ugly code for now, but here's the idea. Before we had only excitatory neurons and tried to model the input solely by current. This was not realistic, since the current was supplied only to neurons corresponding to +1 -> unrealistic. Now, both exc/inh neurons are present based on the input_vec.